Build a generic representing the parameterization of the body of an interface by its implicit `Self` parameter.
This is distinct from the generic representing the interface itself,
which may or may not exist but doesn't have the `Self` parameter.

For formatting purposes, I've merged the inner generic into the
interface definition rather than adding a separate nested `generic`
block. That seemed like the most logical option given that an interface
definition is always generic and doesn't have parameters of its own, and
this avoids an extra level of indentation for every interface. I'm also
using the same SemIR scope name for the two generics and the interface
itself, to minimize scope qualifications.